+++ vehicle-python/src/vehicle_lang/__init__.py
@@ -1,29 +1,48 @@
 from typing import List
 
 from ._version import VERSION as VERSION
+from .check import check as check
+from .compile import compile_to_query as compile_to_query
 from .compile.error import VehicleBuiltinUnsupported as VehicleBuiltinUnsupported
 from .compile.error import VehiclePropertyNotFound as VehiclePropertyNotFound
 
 # from .compile.python import load_loss_function as load_loss_function
 from .error import VehicleError as VehicleError
 from .error import VehicleInternalError as VehicleInternalError
+from .export import export_to_solver as export_to_solver
+from .list import list_properties as list_properties
+from .list import list_resources as list_resources
 from .session.error import VehicleSessionClosed as VehicleSessionClosed
 from .session.error import VehicleSessionUsed as VehicleSessionUsed
 from .typing import AnyOptimiser as AnyOptimiser
 from .typing import AnyOptimisers as AnyOptimisers
 from .typing import DeclarationName as DeclarationName
 from .typing import DifferentiableLogic as DifferentiableLogic
+from .typing import ExportTarget as ExportTarget
 from .typing import Optimiser as Optimiser
 from .typing import QuantifiedVariableName as QuantifiedVariableName
+from .typing import QueryFormat as QueryFormat
+from .typing import TypeSystem as TypeSystem
 from .typing import Verifier as Verifier
+from .validate import validate as validate
 from .verify import verify as verify
 
 __all__: List[str] = [
     "VERSION",
+    # Check
+    "check",
     # Compile
     "load_loss_function",
+    "compile_to_query",
     # Verify
     "verify",
+    # Validate,
+    "validate",
+    # Export
+    "export_to_solver",
+    # List
+    "list_resources",
+    "list_properties",
     # Error types
     "VehicleError",
     "VehicleSessionClosed",
@@ -38,5 +57,8 @@ __all__: List[str] = [
     "AnyOptimiser",
     "AnyOptimisers",
     "DifferentiableLogic",
+    "QueryFormat",
     "Verifier",
+    "TypeSystem",
+    "ExportTarget",
 ]