+++ vehicle/src/Vehicle/Backend/Prelude.hs
@@ -54,6 +54,14 @@ instance Read SecondaryTypeSystem where
     _ -> []
 
 --------------------------------------------------------------------------------
+-- Different Listable Entities
+
+data ListableEntities
+  = ExternalResources
+  | Properties
+  deriving (Eq, Show, Bounded, Enum)
+
+--------------------------------------------------------------------------------
 -- Action
 
 data Target