+++ vehicle/src/Vehicle/CommandLine.hs
@@ -1,3 +1,6 @@
+{-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
+
+{-# HLINT ignore "Use fewer imports" #-}
 module Vehicle.CommandLine
   ( Options (..),
     GlobalOptions (..),
@@ -42,9 +45,10 @@ import Options.Applicative
     switch,
     value,
   )
-import Vehicle.Backend.Prelude (DifferentiableLogicID, ITP, SecondaryTypeSystem (..), Target (..), findTarget)
+import Vehicle.Backend.Prelude (DifferentiableLogicID, ITP, ListableEntities (..), SecondaryTypeSystem (..), Target (..), findTarget)
 import Vehicle.Compile (CompileOptions (..))
 import Vehicle.Export (ExportOptions (..))
+import Vehicle.List (ListOptions (..))
 import Vehicle.Prelude
   ( Doc,
     Pretty (..),
@@ -96,6 +100,7 @@ data ModeOptions
   | Verify VerifyOptions
   | Validate ValidateOptions
   | Export ExportOptions
+  | List ListOptions
   deriving (Eq, Show)
 
 --------------------------------------------------------------------------------
@@ -169,6 +174,7 @@ modeOptionsParser =
         <> command "verify" verifyParserInfo
         <> command "validate" validateParserInfo
         <> command "export" exportParserInfo
+        <> command "list" listParserInfo
 
 --------------------------------------------------------------------------------
 -- Check mode
@@ -189,6 +195,27 @@ typeCheckParser =
 
 typeCheckParserInfo :: ParserInfo ModeOptions
 typeCheckParserInfo = info (Check <$> typeCheckParser) typeCheckDescription
+
+--------------------------------------------------------------------------------
+-- List mode
+
+listDescription :: InfoMod ModeOptions
+listDescription =
+  progDesc $
+    "List entities for a "
+      <> specificationFileExtension
+      <> " specification file"
+      <> "."
+
+listParser :: Parser ListOptions
+listParser =
+  ListOptions
+    <$> listModeParser
+    <*> specificationParser
+    <*> outputAsJSONParser
+
+listParserInfo :: ParserInfo ModeOptions
+listParserInfo = info (List <$> listParser) listDescription
 
 --------------------------------------------------------------------------------
 -- Compile mode
@@ -379,6 +406,22 @@ typeSystemParser =
               )
         )
       <> value Nothing
+
+listModeParser :: Parser ListableEntities
+listModeParser =
+  hsubparser $
+    command
+      "resources"
+      ( info
+          (pure ExternalResources)
+          (progDesc "List all networks, datasets, and parameters in the specification.")
+      )
+      <> command
+        "properties"
+        ( info
+            (pure Properties)
+            (progDesc "List all properties in the specification.")
+        )
 
 specificationParser :: Parser FilePath
 specificationParser =